refactor: set a separate folder for the files related to course controller
